Move contents of control::util module into control::destination::client
These types are only used inside that module anyway.
