+++ src/control/mod.rs
@@ -4,7 +4,6 @@ mod fully_qualified_authority;
 mod observe;
 pub mod pb;
 mod remote_stream;
-mod util;
 
 pub use self::destination::Bind;
 pub use self::observe::Observe;
